🔥 remove unused utils & move evm utils temporarily
